requirements.txt
langchain_community
langchain
streamlit
langchain_core
langserve
fastapi
uvicorn
python-dotenv
sse_starlette
